+++ voxygen/src/anim/bird_medium/mod.rs
@@ -110,6 +110,7 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Eagle, _) => (2.5, 5.0),
                 (Snowyowl, _) => (2.5, 5.0),
                 (Parrot, _) => (0.5, 4.5),
+                (Cockatrice, _) => (3.0, 4.0),
             },
             chest: match (body.species, body.body_type) {
                 (Duck, _) => (0.0, 5.0),
@@ -119,6 +120,7 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Eagle, _) => (0.0, 8.0),
                 (Snowyowl, _) => (0.0, 4.5),
                 (Parrot, _) => (0.0, 5.0),
+                (Cockatrice, _) => (0.0, 12.5),
             },
             tail: match (body.species, body.body_type) {
                 (Duck, _) => (-3.0, 1.5),
@@ -128,6 +130,7 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Eagle, _) => (-8.0, -4.0),
                 (Snowyowl, _) => (-6.0, -2.0),
                 (Parrot, _) => (-8.0, -2.0),
+                (Cockatrice, _) => (-10.0, -2.5),
             },
             wing: match (body.species, body.body_type) {
                 (Duck, _) => (2.75, 0.0, 1.0),
@@ -137,6 +140,7 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Eagle, _) => (3.0, -8.0, 4.0),
                 (Snowyowl, _) => (3.5, -5.5, 4.0),
                 (Parrot, _) => (2.0, -4.5, 3.0),
+                (Cockatrice, _) => (4.5, -2.5, 1.5),
             },
             foot: match (body.species, body.body_type) {
                 (Duck, _) => (2.0, -1.5, 4.0),
@@ -146,6 +150,7 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Eagle, _) => (2.0, -2.0, 8.0),
                 (Snowyowl, _) => (1.5, -2.5, 7.0),
                 (Parrot, _) => (1.5, -3.0, 3.0),
+                (Cockatrice, _) => (4.0, -3.5, 12.0),
             },
         }
     }